ci: allow issues to go to core or devops
In some cases tickets on the vegacapsule will be known to be done by the devops team. By adding the `devops` label when creating these will go directly to the devops project board. All other times the issues will land on the core project board.

This is just to minimise the manual issue movement when we know up-front that the vegacapsule issue is to be worked on by devops